Weather report By city
Added a method which gathers the weather for the city being searched for by the user. The user would have to enter City State and Country. I incoorprated Google Geocoding API to gather the latitude and long of the city the user is requesting weather information for. I than shuffle though the weather stations OpenWeatherMap supports and find the closest weather station to the city the user searched for. I created a meathod using the Haversine formula which calculates the distance btween 2 coordinates as my method of finding the closest weather station.