JavaScript for google map view
Made progress adding the javascript google maps API.  I was able to capture coords of a marker the user would place on the map and send it though the the controller to evaluate the weather conditions. Added the the weather proiperties to the class to properly capture all weather information.